// react-vite-khoaluan/src/types/backend.d.ts
export interface IModelPaginate<T> {
  meta: {
    page: number;
    pageSize: number;
    pages: number;
    total: number;
    hasMore?: boolean;
  };
  result: T[];
}
// ...
export interface ICompany {
  id?: string;
  name?: string;
  address?: string;
  logo?: string | null | undefined;
  description?: string;
  createdBy?: string;
  isDeleted?: boolean;
  deletedAt?: boolean | null;
  createdAt?: string;
  updatedAt?: string;
  field?: string;
  website?: string;
  scale?: string;
  country?: string;
  foundingYear?: number;
  location?: string;
  updatedBy?: string | null;
  totalJobs?: number;
  hrCompany?: hrCompany;
  averageRating?: number;
  totalComments?: number;
  comment?: boolean;
}

export interface hrCompany {
  id?: number;
  name?: string;
  email?: string;
}
// ...
export interface ICommentUser {
  id: number;
  name: string;
  email: string;
  avatar?: string;
}

export interface IComment {
  id: number;
  comment: string;
  rating: number;
  user: ICommentUser;
  createdAt: string;
  createdBy: string;
}